H.W

+++ 03 - Programming Using C++ - Level 1/Input Detailed Information Card H.W/Input Detailed Information Card H.W.cpp
@@ -4,7 +4,7 @@ using namespace std;
 int main()
 {
     string Name;
-    int Age;
+    short Age;
     string Country;
     string City;
     float MonthlySalary;
